Create dynamic array.cpp
Question: At first in the main function take an integer N as input. Then make a function named get_array() which will receive that N as a parameter. Then inside the get_array() function create an integer array of size N. Then the values of that array will be taken as input. After that return that array from that function and receive it in the main function and print the values of the array there.



Sample Input
Sample Output
5
1 2 3 4 5
1 2 3 4 5
7
5 1 2 4 6 9 3
5 1 2 4 6 9 3